requirements.txt
streamlit
SpeechRecognition
pyttsx3
ollama
langchain_ollama
opencv-python
pillow
selenium
numpy
sounddevice